Download the friends and followers in a period of time

This download a time series with the evolution of friends, followers, listed, statuses, likes and the created date of the account

get_friends_followers(screen.name)

Arguments

screen.name
A character vector with an Twitter user

Value

This function returns a data.frame including columns:

date

followers_count

friends_count

listed_count

statuses_count

favourites_count

account_created_at

See also

Examples

get_friends_followers(screen.name = "SergioMassa")
#>  Found 1000 records... Found 1020 records... Imported 1020 records. Simplifying into dataframe...
#>  Found 124 records... Imported 124 records. Simplifying into dataframe...
#>     screen_name       date followers_count friends_count listed_count
#> 1   SergioMassa 2020-12-25         1144207          2075         2289
#> 2   SergioMassa 2021-01-06         1144734          2071         2289
#> 3   SergioMassa 2021-01-23         1145578          2076         2306
#> 4   SergioMassa 2021-01-24         1145591          2076         2305
#> 5   SergioMassa 2021-01-25         1145637          2076         2304
#> 6   SergioMassa 2021-02-11         1146186          2076         2314
#> 7   SergioMassa 2021-02-12         1146194          2075         2315
#> 8   SergioMassa 2021-02-13         1146254          2076         2315
#> 9   SergioMassa 2021-02-14         1146260          2076         2316
#> 10  SergioMassa 2021-02-15         1146327          2076         2316
#> 11  SergioMassa 2021-02-16         1146369          2076         2315
#> 12  SergioMassa 2021-02-17         1146421          2076         2317
#> 13  SergioMassa 2021-02-18         1146469          2076         2316
#> 14  SergioMassa 2021-02-19         1146528          2076         2314
#> 15  SergioMassa 2021-02-20         1146622          2076         2311
#> 16  SergioMassa 2021-02-22         1146824          2077         2310
#> 17  SergioMassa 2021-02-24         1147059          2076         2316
#> 18  SergioMassa 2021-02-25         1147120          2076         2311
#> 19  SergioMassa 2021-02-26         1147170          2077         2312
#> 20  SergioMassa 2021-02-27         1147156          2077         2311
#> 21  SergioMassa 2021-02-28         1147179          2077         2311
#> 22  SergioMassa 2021-03-01         1147264          2077         2311
#> 23  SergioMassa 2021-03-02         1147319          2077            1
#> 24  SergioMassa 2021-03-03         1147382          2077         2318
#> 25  SergioMassa 2021-03-04         1147407          2077         2315
#> 26  SergioMassa 2021-03-05         1147471          2078         2316
#> 27  SergioMassa 2021-03-06         1147530          2078         2318
#> 28  SergioMassa 2021-03-07         1147570          2078         2315
#> 29  SergioMassa 2021-03-08         1147607          2078         2316
#> 30  SergioMassa 2021-03-09         1147565          2078         2314
#> 31  SergioMassa 2021-03-10         1147589          2077         2313
#> 32  SergioMassa 2021-03-11         1147623          2077         2314
#> 33  SergioMassa 2021-03-12         1147634          2077         2314
#> 34  SergioMassa 2021-03-13         1147678          2077         2315
#> 35  SergioMassa 2021-03-14         1147761          2077         2317
#> 36  SergioMassa 2021-03-15         1147791          2077         2317
#> 37  SergioMassa 2021-03-16         1147799          2077         2319
#> 38  SergioMassa 2021-03-17         1147743          2077         2319
#> 39  SergioMassa 2021-03-18         1147775          2076         2320
#> 40  SergioMassa 2021-03-19         1147832          2077         2319
#> 41  SergioMassa 2021-03-20         1147862          2077         2320
#> 42  SergioMassa 2021-03-21         1147824          2077         2321
#> 43  SergioMassa 2021-03-22         1147845          2077         2322
#> 44  SergioMassa 2021-03-23         1147902          2077         2322
#> 45  SergioMassa 2021-03-24         1147919          2077         2323
#> 46  SergioMassa 2021-03-25         1147981          2077         2324
#> 47  SergioMassa 2021-03-26         1148071          2077         2324
#> 48  SergioMassa 2021-03-27         1148089          2077         2325
#> 49  SergioMassa 2021-03-28         1148262          2077         2326
#> 50  SergioMassa 2021-03-29         1148402          2077         2324
#> 51  SergioMassa 2021-03-30         1148451          2078         2327
#> 52  SergioMassa 2021-03-31         1148404          2078         2327
#> 53  SergioMassa 2021-04-01         1148483          2079         2327
#> 54  SergioMassa 2021-04-02         1148532          2079         2327
#> 55  SergioMassa 2021-04-03         1148603          2080         2328
#> 56  SergioMassa 2021-04-04         1148626          2080         2327
#> 57  SergioMassa 2021-04-05         1148668          2080         2327
#> 58  SergioMassa 2021-04-06         1148765          2080         2331
#> 59  SergioMassa 2021-04-07         1148785          2080         2333
#> 60  SergioMassa 2021-04-08         1148677          2080         2331
#> 61  SergioMassa 2021-04-09         1148565          2080         2332
#> 62  SergioMassa 2021-04-10         1148518          2080         2331
#> 63  SergioMassa 2021-04-11         1148483          2079         2330
#> 64  SergioMassa 2021-04-12         1148573          2079         2328
#> 65  SergioMassa 2021-04-13         1148640          2080         2329
#> 66  SergioMassa 2021-04-14         1148744          2080         2329
#> 67  SergioMassa 2021-04-15         1148869          2080         2330
#> 68  SergioMassa 2021-04-16         1149064          2080         2335
#> 69  SergioMassa 2021-04-17         1149086          2080         2333
#> 70  SergioMassa 2021-04-18         1149222          2080         2331
#> 71  SergioMassa 2021-04-19         1149347          2080         2334
#> 72  SergioMassa 2021-04-20         1149464          2080         2336
#> 73  SergioMassa 2021-04-21         1149672          2080         2337
#> 74  SergioMassa 2021-04-22         1149772          2080         2334
#> 75  SergioMassa 2021-04-23         1149854          2080         2335
#> 76  SergioMassa 2021-04-24         1150138          2080         2338
#> 77  SergioMassa 2021-04-28         1150792          2080         2340
#> 78  SergioMassa 2021-04-29         1150907          2080         2338
#> 79  SergioMassa 2021-04-30         1150983          2080         2339
#> 80  SergioMassa 2021-05-01         1151013          2080         2338
#> 81  SergioMassa 2021-05-02         1151064          2080         2339
#> 82  SergioMassa 2021-05-04         1151195          2080         2343
#> 83  SergioMassa 2021-05-05         1151304          2079         2346
#> 84  SergioMassa 2021-05-06         1151421          2079         2344
#> 85  SergioMassa 2021-05-07         1151486          2079         2342
#> 86  SergioMassa 2021-05-08         1151564          2079         2345
#> 87  SergioMassa 2021-05-09         1151560          2078         2344
#> 88  SergioMassa 2021-05-10         1151626          2078         2344
#> 89  SergioMassa 2021-05-11         1151719          2078         2345
#> 90  SergioMassa 2021-05-12         1151742          2079         2344
#> 91  SergioMassa 2021-05-13         1151617          2079         2344
#> 92  SergioMassa 2021-05-14         1151449          2079         2345
#> 93  SergioMassa 2021-05-15         1151230          2078         2345
#> 94  SergioMassa 2021-05-16         1151041          2078         2345
#> 95  SergioMassa 2021-05-17         1150688          2078         2344
#> 96  SergioMassa 2021-05-18         1150372          2077         2346
#> 97  SergioMassa 2021-05-19         1149956          2077         2347
#> 98  SergioMassa 2021-05-20         1149591          2077         2346
#> 99  SergioMassa 2021-05-21         1149226          2077         2344
#> 100 SergioMassa 2021-05-22         1148972          2077         2344
#> 101 SergioMassa 2021-05-23         1149000          2077         2343
#> 102 SergioMassa 2021-05-24         1149039          2078         2344
#> 103 SergioMassa 2021-05-25         1149102          2078         2344
#> 104 SergioMassa 2021-05-26         1148943          2078         2345
#> 105 SergioMassa 2021-05-27         1148647          2079         2345
#> 106 SergioMassa 2021-05-28         1148319          2079         2343
#> 107 SergioMassa 2021-05-29         1148042          2079         2344
#> 108 SergioMassa 2021-06-01         1147677          2080         2344
#> 109 SergioMassa 2021-06-04         1147833          2080         2345
#> 110 SergioMassa 2021-06-05         1147868          2080         2344
#> 111 SergioMassa 2021-06-06         1147933          2080         2345
#> 112 SergioMassa 2021-06-07         1147973          2080         2345
#> 113 SergioMassa 2021-06-08         1148021          2080         2345
#> 114 SergioMassa 2021-06-09         1147886          2080         2344
#> 115 SergioMassa 2021-06-10         1147176          2080         2345
#> 116 SergioMassa 2021-06-11         1147050          2080         2345
#> 117 SergioMassa 2021-06-12         1147083          2080         2347
#> 118 SergioMassa 2021-06-13         1147134          2094         2346
#> 119 SergioMassa 2021-06-14         1147179          2094         2346
#> 120 SergioMassa 2021-06-16         1147262          2093         2346
#> 121 SergioMassa 2021-06-17         1147340          2092         2348
#> 122 SergioMassa 2021-06-18         1147427          2093         2348
#> 123 SergioMassa 2021-06-22         1147618          2093         2350
#> 124 SergioMassa 2021-06-25         1147827          2093         2348
#>     statuses_count favourites_count
#> 1             9069            11097
#> 2             9075            11102
#> 3             9082            11106
#> 4             9082            11106
#> 5             9082            11106
#> 6             9092            11113
#> 7             9099            11117
#> 8             9098            11115
#> 9             9098            11115
#> 10            9099            11115
#> 11            9099            11115
#> 12            9099            11114
#> 13            9101            11116
#> 14            9102            11117
#> 15            9102            11117
#> 16            9102            11117
#> 17            9102            11117
#> 18            9102            11117
#> 19            9102            11117
#> 20            9102            11117
#> 21            9102            11117
#> 22            9103            11117
#> 23            9104            11118
#> 24            9104            11118
#> 25            9104            11118
#> 26            9104            11118
#> 27            9108            11119
#> 28            9108            11119
#> 29            9108            11118
#> 30            9109            11124
#> 31            9110            11125
#> 32            9110            11125
#> 33            9113            11125
#> 34            9114            11125
#> 35            9115            11126
#> 36            9115            11126
#> 37            9119            11125
#> 38            9121            11127
#> 39            9122            11128
#> 40            9122            11128
#> 41            9122            11128
#> 42            9122            11128
#> 43            9122            11128
#> 44            9122            11128
#> 45            9122            11128
#> 46            9123            11129
#> 47            9125            11131
#> 48            9125            11131
#> 49            9130            11135
#> 50            9131            11135
#> 51            9131            11136
#> 52            9134            11138
#> 53            9134            11137
#> 54            9134            11137
#> 55            9136            11139
#> 56            9136            11138
#> 57            9136            11138
#> 58            9136            11138
#> 59            9138            11140
#> 60            9140            11140
#> 61            9140            11140
#> 62            9142            11141
#> 63            9142            11141
#> 64            9143            11141
#> 65            9143            11141
#> 66            9143            11143
#> 67            9144            11144
#> 68            9144            11141
#> 69            9145            11143
#> 70            9145            11143
#> 71            9145            11143
#> 72            9145            11143
#> 73            9153            11144
#> 74            9153            11144
#> 75            9156            11147
#> 76            9158            11148
#> 77            9161            11149
#> 78            9161            11149
#> 79            9161            11149
#> 80            9161            11149
#> 81            9162            11150
#> 82            9163            11149
#> 83            9163            11149
#> 84            9164            11150
#> 85            9164            11150
#> 86            9168            11154
#> 87            9168            11154
#> 88            9172            11157
#> 89            9172            11158
#> 90            9173            11158
#> 91            9174            11159
#> 92            9176            11161
#> 93            9177            11162
#> 94            9177            11162
#> 95            9177            11162
#> 96            9179            11164
#> 97            9179            11164
#> 98            9185            11169
#> 99            9185            11169
#> 100           9185            11169
#> 101           9185            11169
#> 102           9185            11172
#> 103           9185            11172
#> 104           9188            11174
#> 105           9188            11176
#> 106           9190            11176
#> 107           9190            11176
#> 108           9190            11176
#> 109           9193            11183
#> 110           9199            11187
#> 111           9202            11190
#> 112           9202            11190
#> 113           9202            11190
#> 114           9202            11190
#> 115           9203            11191
#> 116           9208            11196
#> 117           9210            11196
#> 118           9210            11196
#> 119           9210            11196
#> 120           9214            11198
#> 121           9218            11201
#> 122           9221            11201
#> 123           9222            11202
#> 124           9228            11204
get_friends_followers(screen.name = c("SergioMassa", "dalesmm"))
#>  Found 1000 records... Found 1020 records... Imported 1020 records. Simplifying into dataframe...
#>  Found 124 records... Imported 124 records. Simplifying into dataframe...
#>  Found 127 records... Imported 127 records. Simplifying into dataframe...
#>     screen_name       date followers_count friends_count listed_count
#> 1   SergioMassa 2020-12-25         1144207          2075         2289
#> 2   SergioMassa 2021-01-06         1144734          2071         2289
#> 3   SergioMassa 2021-01-23         1145578          2076         2306
#> 4   SergioMassa 2021-01-24         1145591          2076         2305
#> 5   SergioMassa 2021-01-25         1145637          2076         2304
#> 6   SergioMassa 2021-02-11         1146186          2076         2314
#> 7   SergioMassa 2021-02-12         1146194          2075         2315
#> 8   SergioMassa 2021-02-13         1146254          2076         2315
#> 9   SergioMassa 2021-02-14         1146260          2076         2316
#> 10  SergioMassa 2021-02-15         1146327          2076         2316
#> 11  SergioMassa 2021-02-16         1146369          2076         2315
#> 12  SergioMassa 2021-02-17         1146421          2076         2317
#> 13  SergioMassa 2021-02-18         1146469          2076         2316
#> 14  SergioMassa 2021-02-19         1146528          2076         2314
#> 15  SergioMassa 2021-02-20         1146622          2076         2311
#> 16  SergioMassa 2021-02-22         1146824          2077         2310
#> 17  SergioMassa 2021-02-24         1147059          2076         2316
#> 18  SergioMassa 2021-02-25         1147120          2076         2311
#> 19  SergioMassa 2021-02-26         1147170          2077         2312
#> 20  SergioMassa 2021-02-27         1147156          2077         2311
#> 21  SergioMassa 2021-02-28         1147179          2077         2311
#> 22  SergioMassa 2021-03-01         1147264          2077         2311
#> 23  SergioMassa 2021-03-02         1147319          2077            1
#> 24  SergioMassa 2021-03-03         1147382          2077         2318
#> 25  SergioMassa 2021-03-04         1147407          2077         2315
#> 26  SergioMassa 2021-03-05         1147471          2078         2316
#> 27  SergioMassa 2021-03-06         1147530          2078         2318
#> 28  SergioMassa 2021-03-07         1147570          2078         2315
#> 29  SergioMassa 2021-03-08         1147607          2078         2316
#> 30  SergioMassa 2021-03-09         1147565          2078         2314
#> 31  SergioMassa 2021-03-10         1147589          2077         2313
#> 32  SergioMassa 2021-03-11         1147623          2077         2314
#> 33  SergioMassa 2021-03-12         1147634          2077         2314
#> 34  SergioMassa 2021-03-13         1147678          2077         2315
#> 35  SergioMassa 2021-03-14         1147761          2077         2317
#> 36  SergioMassa 2021-03-15         1147791          2077         2317
#> 37  SergioMassa 2021-03-16         1147799          2077         2319
#> 38  SergioMassa 2021-03-17         1147743          2077         2319
#> 39  SergioMassa 2021-03-18         1147775          2076         2320
#> 40  SergioMassa 2021-03-19         1147832          2077         2319
#> 41  SergioMassa 2021-03-20         1147862          2077         2320
#> 42  SergioMassa 2021-03-21         1147824          2077         2321
#> 43  SergioMassa 2021-03-22         1147845          2077         2322
#> 44  SergioMassa 2021-03-23         1147902          2077         2322
#> 45  SergioMassa 2021-03-24         1147919          2077         2323
#> 46  SergioMassa 2021-03-25         1147981          2077         2324
#> 47  SergioMassa 2021-03-26         1148071          2077         2324
#> 48  SergioMassa 2021-03-27         1148089          2077         2325
#> 49  SergioMassa 2021-03-28         1148262          2077         2326
#> 50  SergioMassa 2021-03-29         1148402          2077         2324
#> 51  SergioMassa 2021-03-30         1148451          2078         2327
#> 52  SergioMassa 2021-03-31         1148404          2078         2327
#> 53  SergioMassa 2021-04-01         1148483          2079         2327
#> 54  SergioMassa 2021-04-02         1148532          2079         2327
#> 55  SergioMassa 2021-04-03         1148603          2080         2328
#> 56  SergioMassa 2021-04-04         1148626          2080         2327
#> 57  SergioMassa 2021-04-05         1148668          2080         2327
#> 58  SergioMassa 2021-04-06         1148765          2080         2331
#> 59  SergioMassa 2021-04-07         1148785          2080         2333
#> 60  SergioMassa 2021-04-08         1148677          2080         2331
#> 61  SergioMassa 2021-04-09         1148565          2080         2332
#> 62  SergioMassa 2021-04-10         1148518          2080         2331
#> 63  SergioMassa 2021-04-11         1148483          2079         2330
#> 64  SergioMassa 2021-04-12         1148573          2079         2328
#> 65  SergioMassa 2021-04-13         1148640          2080         2329
#> 66  SergioMassa 2021-04-14         1148744          2080         2329
#> 67  SergioMassa 2021-04-15         1148869          2080         2330
#> 68  SergioMassa 2021-04-16         1149064          2080         2335
#> 69  SergioMassa 2021-04-17         1149086          2080         2333
#> 70  SergioMassa 2021-04-18         1149222          2080         2331
#> 71  SergioMassa 2021-04-19         1149347          2080         2334
#> 72  SergioMassa 2021-04-20         1149464          2080         2336
#> 73  SergioMassa 2021-04-21         1149672          2080         2337
#> 74  SergioMassa 2021-04-22         1149772          2080         2334
#> 75  SergioMassa 2021-04-23         1149854          2080         2335
#> 76  SergioMassa 2021-04-24         1150138          2080         2338
#> 77  SergioMassa 2021-04-28         1150792          2080         2340
#> 78  SergioMassa 2021-04-29         1150907          2080         2338
#> 79  SergioMassa 2021-04-30         1150983          2080         2339
#> 80  SergioMassa 2021-05-01         1151013          2080         2338
#> 81  SergioMassa 2021-05-02         1151064          2080         2339
#> 82  SergioMassa 2021-05-04         1151195          2080         2343
#> 83  SergioMassa 2021-05-05         1151304          2079         2346
#> 84  SergioMassa 2021-05-06         1151421          2079         2344
#> 85  SergioMassa 2021-05-07         1151486          2079         2342
#> 86  SergioMassa 2021-05-08         1151564          2079         2345
#> 87  SergioMassa 2021-05-09         1151560          2078         2344
#> 88  SergioMassa 2021-05-10         1151626          2078         2344
#> 89  SergioMassa 2021-05-11         1151719          2078         2345
#> 90  SergioMassa 2021-05-12         1151742          2079         2344
#> 91  SergioMassa 2021-05-13         1151617          2079         2344
#> 92  SergioMassa 2021-05-14         1151449          2079         2345
#> 93  SergioMassa 2021-05-15         1151230          2078         2345
#> 94  SergioMassa 2021-05-16         1151041          2078         2345
#> 95  SergioMassa 2021-05-17         1150688          2078         2344
#> 96  SergioMassa 2021-05-18         1150372          2077         2346
#> 97  SergioMassa 2021-05-19         1149956          2077         2347
#> 98  SergioMassa 2021-05-20         1149591          2077         2346
#> 99  SergioMassa 2021-05-21         1149226          2077         2344
#> 100 SergioMassa 2021-05-22         1148972          2077         2344
#> 101 SergioMassa 2021-05-23         1149000          2077         2343
#> 102 SergioMassa 2021-05-24         1149039          2078         2344
#> 103 SergioMassa 2021-05-25         1149102          2078         2344
#> 104 SergioMassa 2021-05-26         1148943          2078         2345
#> 105 SergioMassa 2021-05-27         1148647          2079         2345
#> 106 SergioMassa 2021-05-28         1148319          2079         2343
#> 107 SergioMassa 2021-05-29         1148042          2079         2344
#> 108 SergioMassa 2021-06-01         1147677          2080         2344
#> 109 SergioMassa 2021-06-04         1147833          2080         2345
#> 110 SergioMassa 2021-06-05         1147868          2080         2344
#> 111 SergioMassa 2021-06-06         1147933          2080         2345
#> 112 SergioMassa 2021-06-07         1147973          2080         2345
#> 113 SergioMassa 2021-06-08         1148021          2080         2345
#> 114 SergioMassa 2021-06-09         1147886          2080         2344
#> 115 SergioMassa 2021-06-10         1147176          2080         2345
#> 116 SergioMassa 2021-06-11         1147050          2080         2345
#> 117 SergioMassa 2021-06-12         1147083          2080         2347
#> 118 SergioMassa 2021-06-13         1147134          2094         2346
#> 119 SergioMassa 2021-06-14         1147179          2094         2346
#> 120 SergioMassa 2021-06-16         1147262          2093         2346
#> 121 SergioMassa 2021-06-17         1147340          2092         2348
#> 122 SergioMassa 2021-06-18         1147427          2093         2348
#> 123 SergioMassa 2021-06-22         1147618          2093         2350
#> 124 SergioMassa 2021-06-25         1147827          2093         2348
#> 125     dalesmm 2020-12-25           61709          5701          288
#> 126     dalesmm 2021-01-06           61991          5701          289
#> 127     dalesmm 2021-01-23           62336          5722          289
#> 128     dalesmm 2021-01-24           62342          5722          289
#> 129     dalesmm 2021-01-25           62350          5723          289
#> 130     dalesmm 2021-02-01           62407          5735          288
#> 131     dalesmm 2021-02-02           62408          5738          288
#> 132     dalesmm 2021-02-03           62427          5740          288
#> 133     dalesmm 2021-02-11           62453          5751          289
#> 134     dalesmm 2021-02-12           62446          5755          289
#> 135     dalesmm 2021-02-13           62456          5756          289
#> 136     dalesmm 2021-02-14           62460          5756          289
#> 137     dalesmm 2021-02-15           62464          5762          289
#> 138     dalesmm 2021-02-16           62467          5761          289
#> 139     dalesmm 2021-02-17           62470          5759          289
#> 140     dalesmm 2021-02-18           63233          5764          291
#> 141     dalesmm 2021-02-19           63413          5765          291
#> 142     dalesmm 2021-02-20           63486          5767          292
#> 143     dalesmm 2021-02-22           65392          5782          292
#> 144     dalesmm 2021-02-24           65665          5783          293
#> 145     dalesmm 2021-02-25           65698          5783          293
#> 146     dalesmm 2021-02-26           65722          5783          293
#> 147     dalesmm 2021-02-27           65730          5783          294
#> 148     dalesmm 2021-02-28           65754          5784          295
#> 149     dalesmm 2021-03-01           65785          5790          294
#> 150     dalesmm 2021-03-02           65820          5789          295
#> 151     dalesmm 2021-03-03           65833          5789          295
#> 152     dalesmm 2021-03-04           65842          5790          294
#> 153     dalesmm 2021-03-05           65862          5791          294
#> 154     dalesmm 2021-03-06           65880          5792          297
#> 155     dalesmm 2021-03-07           65892          5793          295
#> 156     dalesmm 2021-03-08           65918          5798          295
#> 157     dalesmm 2021-03-09           65974          5803          296
#> 158     dalesmm 2021-03-10           65988          5807          296
#> 159     dalesmm 2021-03-11           65994          5809          297
#> 160     dalesmm 2021-03-12           65994          5809          297
#> 161     dalesmm 2021-03-13           66005          5809          297
#> 162     dalesmm 2021-03-14           66008          5809          297
#> 163     dalesmm 2021-03-15           66004          5810          297
#> 164     dalesmm 2021-03-16           66010          5813          297
#> 165     dalesmm 2021-03-17           66017          5817          296
#> 166     dalesmm 2021-03-18           66024          5823          296
#> 167     dalesmm 2021-03-19           66026          5823          295
#> 168     dalesmm 2021-03-20           66043          5825          295
#> 169     dalesmm 2021-03-21           66041          5826          295
#> 170     dalesmm 2021-03-22           66042          5826          296
#> 171     dalesmm 2021-03-23           66038          5829          295
#> 172     dalesmm 2021-03-24           66045          5831          295
#> 173     dalesmm 2021-03-25           66053          5833          295
#> 174     dalesmm 2021-03-26           66057          5835          295
#> 175     dalesmm 2021-03-27           66065          5835          295
#> 176     dalesmm 2021-03-28           66065          5837          296
#> 177     dalesmm 2021-03-29           66068          5841          295
#> 178     dalesmm 2021-03-30           66079          5849          295
#> 179     dalesmm 2021-03-31           66088          5853          295
#> 180     dalesmm 2021-04-01           66090          5853          295
#> 181     dalesmm 2021-04-02           66107          5855          295
#> 182     dalesmm 2021-04-03           66106          5855          295
#> 183     dalesmm 2021-04-04           66102          5855          295
#> 184     dalesmm 2021-04-05           66099          5857          295
#> 185     dalesmm 2021-04-06           66106          5857          295
#> 186     dalesmm 2021-04-07           66103          5858          295
#> 187     dalesmm 2021-04-08           66092          5862          295
#> 188     dalesmm 2021-04-09           66095          5866          295
#> 189     dalesmm 2021-04-10           66109          5868          295
#> 190     dalesmm 2021-04-11           66118          5868          295
#> 191     dalesmm 2021-04-12           66124          5868          295
#> 192     dalesmm 2021-04-13           66128          5867          296
#> 193     dalesmm 2021-04-14           66143          5873          296
#> 194     dalesmm 2021-04-15           66139          5878          295
#> 195     dalesmm 2021-04-16           66137          5878          295
#> 196     dalesmm 2021-04-17           66122          5878          295
#> 197     dalesmm 2021-04-18           66143          5879          295
#> 198     dalesmm 2021-04-19           66142          5880          295
#> 199     dalesmm 2021-04-20           66154          5881          295
#> 200     dalesmm 2021-04-21           66168          5885          298
#> 201     dalesmm 2021-04-22           66165          5884          297
#> 202     dalesmm 2021-04-23           66177          5886          297
#> 203     dalesmm 2021-04-24           66168          5886          297
#> 204     dalesmm 2021-04-28           66193          5899          298
#> 205     dalesmm 2021-04-29           66196          5901          298
#> 206     dalesmm 2021-04-30           66204          5901          298
#> 207     dalesmm 2021-05-01           66216          5901          298
#> 208     dalesmm 2021-05-02           66219          5904          298
#> 209     dalesmm 2021-05-04           66253          5911          299
#> 210     dalesmm 2021-05-05           66267          5913          299
#> 211     dalesmm 2021-05-06           66291          5913          299
#> 212     dalesmm 2021-05-07           66311          5916          299
#> 213     dalesmm 2021-05-08           66319          5917          299
#> 214     dalesmm 2021-05-09           66321          5917          299
#> 215     dalesmm 2021-05-10           66337          5920          299
#> 216     dalesmm 2021-05-11           66349          5925          299
#> 217     dalesmm 2021-05-12           66382          5926          299
#> 218     dalesmm 2021-05-13           66421          5928          299
#> 219     dalesmm 2021-05-14           66503          5930          299
#> 220     dalesmm 2021-05-15           66507          5929          299
#> 221     dalesmm 2021-05-16           66495          5931          299
#> 222     dalesmm 2021-05-17           66495          5934          299
#> 223     dalesmm 2021-05-18           66522          5934          299
#> 224     dalesmm 2021-05-19           66512          5934          301
#> 225     dalesmm 2021-05-20           66518          5937          301
#> 226     dalesmm 2021-05-21           66516          5936          301
#> 227     dalesmm 2021-05-22           66515          5936          301
#> 228     dalesmm 2021-05-23           66523          5939          301
#> 229     dalesmm 2021-05-24           66519          5941          302
#> 230     dalesmm 2021-05-25           66522          5944          302
#> 231     dalesmm 2021-05-26           66518          5945          302
#> 232     dalesmm 2021-05-27           66522          5947            0
#> 233     dalesmm 2021-05-28           66526          5951          302
#> 234     dalesmm 2021-05-29           66558          5952          302
#> 235     dalesmm 2021-06-01           66683          5958          303
#> 236     dalesmm 2021-06-04           66742          5958          304
#> 237     dalesmm 2021-06-05           66759          5958          304
#> 238     dalesmm 2021-06-06           66772          5961          304
#> 239     dalesmm 2021-06-07           66792          5963          304
#> 240     dalesmm 2021-06-08           66806          5964          304
#> 241     dalesmm 2021-06-09           66812          5964          303
#> 242     dalesmm 2021-06-10           66808          5971          302
#> 243     dalesmm 2021-06-11           66802          5972          302
#> 244     dalesmm 2021-06-12           66829          5975          303
#> 245     dalesmm 2021-06-13           66842          5976          303
#> 246     dalesmm 2021-06-14           66865          5969          303
#> 247     dalesmm 2021-06-16           66895          5972          305
#> 248     dalesmm 2021-06-17           66909          5973          305
#> 249     dalesmm 2021-06-18           66920          5973          304
#> 250     dalesmm 2021-06-22           66941          5979          305
#> 251     dalesmm 2021-06-25           66984          5982          305
#>     statuses_count favourites_count
#> 1             9069            11097
#> 2             9075            11102
#> 3             9082            11106
#> 4             9082            11106
#> 5             9082            11106
#> 6             9092            11113
#> 7             9099            11117
#> 8             9098            11115
#> 9             9098            11115
#> 10            9099            11115
#> 11            9099            11115
#> 12            9099            11114
#> 13            9101            11116
#> 14            9102            11117
#> 15            9102            11117
#> 16            9102            11117
#> 17            9102            11117
#> 18            9102            11117
#> 19            9102            11117
#> 20            9102            11117
#> 21            9102            11117
#> 22            9103            11117
#> 23            9104            11118
#> 24            9104            11118
#> 25            9104            11118
#> 26            9104            11118
#> 27            9108            11119
#> 28            9108            11119
#> 29            9108            11118
#> 30            9109            11124
#> 31            9110            11125
#> 32            9110            11125
#> 33            9113            11125
#> 34            9114            11125
#> 35            9115            11126
#> 36            9115            11126
#> 37            9119            11125
#> 38            9121            11127
#> 39            9122            11128
#> 40            9122            11128
#> 41            9122            11128
#> 42            9122            11128
#> 43            9122            11128
#> 44            9122            11128
#> 45            9122            11128
#> 46            9123            11129
#> 47            9125            11131
#> 48            9125            11131
#> 49            9130            11135
#> 50            9131            11135
#> 51            9131            11136
#> 52            9134            11138
#> 53            9134            11137
#> 54            9134            11137
#> 55            9136            11139
#> 56            9136            11138
#> 57            9136            11138
#> 58            9136            11138
#> 59            9138            11140
#> 60            9140            11140
#> 61            9140            11140
#> 62            9142            11141
#> 63            9142            11141
#> 64            9143            11141
#> 65            9143            11141
#> 66            9143            11143
#> 67            9144            11144
#> 68            9144            11141
#> 69            9145            11143
#> 70            9145            11143
#> 71            9145            11143
#> 72            9145            11143
#> 73            9153            11144
#> 74            9153            11144
#> 75            9156            11147
#> 76            9158            11148
#> 77            9161            11149
#> 78            9161            11149
#> 79            9161            11149
#> 80            9161            11149
#> 81            9162            11150
#> 82            9163            11149
#> 83            9163            11149
#> 84            9164            11150
#> 85            9164            11150
#> 86            9168            11154
#> 87            9168            11154
#> 88            9172            11157
#> 89            9172            11158
#> 90            9173            11158
#> 91            9174            11159
#> 92            9176            11161
#> 93            9177            11162
#> 94            9177            11162
#> 95            9177            11162
#> 96            9179            11164
#> 97            9179            11164
#> 98            9185            11169
#> 99            9185            11169
#> 100           9185            11169
#> 101           9185            11169
#> 102           9185            11172
#> 103           9185            11172
#> 104           9188            11174
#> 105           9188            11176
#> 106           9190            11176
#> 107           9190            11176
#> 108           9190            11176
#> 109           9193            11183
#> 110           9199            11187
#> 111           9202            11190
#> 112           9202            11190
#> 113           9202            11190
#> 114           9202            11190
#> 115           9203            11191
#> 116           9208            11196
#> 117           9210            11196
#> 118           9210            11196
#> 119           9210            11196
#> 120           9214            11198
#> 121           9218            11201
#> 122           9221            11201
#> 123           9222            11202
#> 124           9228            11204
#> 125          77559           146926
#> 126          77706           147369
#> 127          77926           147831
#> 128          77927           147844
#> 129          77935           147852
#> 130          78000           148046
#> 131          78005           148058
#> 132          78016           148084
#> 133          78076           148240
#> 134          78092           148269
#> 135          78100           148281
#> 136          78108           148303
#> 137          78108           148305
#> 138          78110           148322
#> 139          78115           148336
#> 140          78213           148612
#> 141          78239           148691
#> 142          78259           148719
#> 143          78296           148862
#> 144          78304           148887
#> 145          78312           148889
#> 146          78326           148898
#> 147          78340           148921
#> 148          78349           148952
#> 149          78381           149047
#> 150          78393           149059
#> 151          78409           149049
#> 152          78420           149070
#> 153          78433           149077
#> 154          78442           149098
#> 155          78453           149083
#> 156          78469           149128
#> 157          78507           149201
#> 158          78534           149253
#> 159          78553           149262
#> 160          78564           149281
#> 161          78571           149268
#> 162          78581           149282
#> 163          78595           149305
#> 164          78606           149329
#> 165          78628           149370
#> 166          78646           149405
#> 167          78650           149412
#> 168          78662           149440
#> 169          78664           149444
#> 170          78665           149449
#> 171          78672           149464
#> 172          78675           149488
#> 173          78685           149518
#> 174          78697           149533
#> 175          78716           149568
#> 176          78718           149577
#> 177          78727           149605
#> 178          78736           149658
#> 179          78741           149683
#> 180          78749           149694
#> 181          78749           149711
#> 182          78752           149718
#> 183          78756           149751
#> 184          78759           149770
#> 185          78762           149776
#> 186          78770           149804
#> 187          78784           149830
#> 188          78798           149852
#> 189          78814           149883
#> 190          78813           149875
#> 191          78816           149885
#> 192          78820           149906
#> 193          78845           149951
#> 194          78873           149998
#> 195          78881           150015
#> 196          78900           150034
#> 197          78913           150061
#> 198          78920           150086
#> 199          78944           150139
#> 200          78961           150187
#> 201          78977           150233
#> 202          78998           150293
#> 203          79006           150306
#> 204          79057           150463
#> 205          79074           150490
#> 206          79082           150498
#> 207          79084           150506
#> 208          79086           150526
#> 209          79115           150598
#> 210          79132           150633
#> 211          79158           150672
#> 212          79184           150681
#> 213          79195           150712
#> 214          79200           150728
#> 215          79226           150771
#> 216          79246           150795
#> 217          79259           150821
#> 218          79291           150902
#> 219          79306           150944
#> 220          79316           150941
#> 221          79326           150974
#> 222          79360           151036
#> 223          79367           151111
#> 224          79381           151137
#> 225          79410           151177
#> 226          79414           151196
#> 227          79425           151210
#> 228          79440           151249
#> 229          79446           151291
#> 230          79453           151308
#> 231          79465           151365
#> 232          79477           151386
#> 233          79507           151443
#> 234          79563           151560
#> 235          79616           151719
#> 236          79674           151842
#> 237          79679           151873
#> 238          79687           151891
#> 239          79692           151911
#> 240          79706           151938
#> 241          79719           151969
#> 242          79737           152010
#> 243          79771           152077
#> 244          79775           152112
#> 245          79780           152138
#> 246          79781           152150
#> 247          79795           152189
#> 248          79803           152206
#> 249          79823           152247
#> 250          79861           152347
#> 251          79880           152395