Download the data of who's retweets who's

This function download all the retweets that every user on the list of politician on twitter made with each others

get_network_data(
  category = "all",
  start_date = "2021-01-01",
  end_date = "2021-03-31"
)

Arguments

category
A character vector with the category selected -"deputies","senators","national executive","others","province servants", "all"-

start_date
A character with the date where the retweets ocurr. This param read the date info in format 'yyyy-mm-dd'

end_date
A character with the date where the retweets stop. This param read the date info in format 'yyyy-mm-dd'

Examples

get_network_data(category = 'province servants',  start_date= "2020-10-01", end_date = "2020-12-31")
#>  Found 1000 records... Found 1020 records... Imported 1020 records. Simplifying into dataframe...
#>  Found 1000 records... Found 2000 records... Found 3000 records... Found 4000 records... Found 5000 records... Found 6000 records... Found 7000 records... Found 8000 records... Found 9000 records... Found 10000 records... Found 11000 records... Found 12000 records... Found 13000 records... Found 14000 records... Found 15000 records... Found 16000 records... Found 17000 records... Found 18000 records... Found 19000 records... Found 20000 records... Found 21000 records... Found 22000 records... Found 23000 records... Found 24000 records... Found 25000 records... Found 26000 records... Found 27000 records... Found 28000 records... Found 29000 records... Found 30000 records... Found 31000 records... Found 32000 records... Found 33000 records... Found 34000 records... Found 35000 records... Found 36000 records... Found 37000 records... Found 38000 records... Found 39000 records... Found 40000 records... Found 41000 records... Found 42000 records... Found 43000 records... Found 44000 records... Found 45000 records... Found 46000 records... Found 47000 records... Found 48000 records... Found 49000 records... Found 50000 records... Found 51000 records... Found 52000 records... Found 53000 records... Found 54000 records... Found 55000 records... Found 56000 records... Found 57000 records... Found 58000 records... Found 59000 records... Found 60000 records... Found 61000 records... Found 62000 records... Found 63000 records... Found 64000 records... Found 65000 records... Found 66000 records... Found 67000 records... Found 68000 records... Found 69000 records... Found 70000 records... Found 71000 records... Found 72000 records... Found 73000 records... Found 74000 records... Found 75000 records... Found 76000 records... Found 77000 records... Found 78000 records... Found 79000 records... Found 80000 records... Found 81000 records... Found 82000 records... Found 83000 records... Found 84000 records... Found 85000 records... Found 86000 records... Found 87000 records... Found 88000 records... Found 89000 records... Found 90000 records... Found 91000 records... Found 92000 records... Found 92848 records... Imported 92848 records. Simplifying into dataframe...
#>       month_year             user_id     retweet_user_id value
#> 45911 2020-10-31           110043455           500964169     1
#> 46101 2020-10-31 1213976332750618624          2953955753     5
#> 46105 2020-10-31 1213976332750618624           500964169     1
#> 46106 2020-10-31 1213976332750618624  779340061263425537     1
#> 46288 2020-10-31          1475504383          2953955753     2
#> 46290 2020-10-31          1475504383           500964169     1
#> 46403 2020-10-31          1579326655           143689068     1
#> 46407 2020-10-31          1579326655          2953955753     3
#> 46416 2020-10-31          1584076484          2953955753     2
#> 46419 2020-10-31          1584076484           500964169     3
#> 46591 2020-10-31           174801590 1239520164699398145     1
#> 46627 2020-10-31           185076776          2953955753    45
#> 46630 2020-10-31           185076776           500964169     5
#> 46632 2020-10-31           185076776  879024609227665408     1
#> 46649 2020-10-31           188998649          2953955753     7
#> 46650 2020-10-31           188998649          4707533116     1
#> 46690 2020-10-31           200859218          1691697295     2
#> 47016 2020-10-31           288342621           172841079     1
#> 47017 2020-10-31           288342621            37494271     2
#> 47029 2020-10-31          2953955753           143689068     1
#> 47034 2020-10-31          2953955753           500964169     7
#> 47036 2020-10-31          2953955753  879024609227665408     2
#> 47037 2020-10-31           297543171            37494271     5
#> 47038 2020-10-31           297543171            77009449     2
#> 47234 2020-10-31            34455479            37494271     3
#> 47235 2020-10-31            34455479            77009449    54
#> 47320 2020-10-31          3763467682           143689068     3
#> 47323 2020-10-31          3763467682          1579326655     3
#> 47324 2020-10-31          3763467682          1584076484     4
#> 47325 2020-10-31          3763467682           185076776     1
#> 47326 2020-10-31          3763467682          2953955753    17
#> 47329 2020-10-31          3763467682           500964169     4
#> 47333 2020-10-31          3763467682  779340061263425537    74
#> 47492 2020-10-31          4707533116           143689068     3
#> 47493 2020-10-31          4707533116          1475504383     1
#> 47496 2020-10-31          4707533116          2953955753     3
#> 47497 2020-10-31          4707533116           500964169     2
#> 47556 2020-10-31           500964169 1215668032934117378     1
#> 47557 2020-10-31           500964169          1579326655     2
#> 47558 2020-10-31           500964169           188998649     1
#> 47559 2020-10-31           500964169          2953955753     1
#> 47562 2020-10-31           500964169  779340061263425537     1
#> 47563 2020-10-31           500964169  879024609227665408     2
#> 48051 2020-10-31           950224302           185076776     1
#> 48052 2020-10-31           950224302          2953955753     4
#> 48259 2020-11-30 1213976332750618624          2953955753     6
#> 48263 2020-11-30 1213976332750618624           500964169     6
#> 48482 2020-11-30          1579326655           143689068     1
#> 48485 2020-11-30          1579326655          2953955753     2
#> 48493 2020-11-30          1584076484           500964169     4
#> 48494 2020-11-30          1584076484  779340061263425537     2
#> 48495 2020-11-30          1584076484           950224302     1
#> 48632 2020-11-30           185076776          2953955753    17
#> 48633 2020-11-30           185076776           500964169     2
#> 48634 2020-11-30           185076776           593781032     1
#> 48636 2020-11-30           188998649          2953955753     4
#> 48655 2020-11-30           200859218          1691697295     1
#> 48786 2020-11-30          2544467503           143852870     1
#> 48897 2020-11-30           288342621           164368516     1
#> 48898 2020-11-30           288342621           172841079     1
#> 48899 2020-11-30           288342621            17449646     1
#> 48900 2020-11-30           288342621           297543171     1
#> 48901 2020-11-30           288342621            37494271     2
#> 48903 2020-11-30           288342621            77009449     2
#> 48910 2020-11-30          2953955753          1579326655     1
#> 48912 2020-11-30          2953955753           500964169     7
#> 48915 2020-11-30          2953955753  879024609227665408     1
#> 48916 2020-11-30           297543171            37494271     5
#> 48917 2020-11-30           297543171            77009449     2
#> 48963 2020-11-30          3256203501          2544467503     2
#> 49032 2020-11-30            34455479            37494271     2
#> 49033 2020-11-30            34455479            77009449    63
#> 49073 2020-11-30          3763467682          1584076484     1
#> 49075 2020-11-30          3763467682          2589783002     1
#> 49076 2020-11-30          3763467682          2953955753     3
#> 49079 2020-11-30          3763467682           500964169     4
#> 49082 2020-11-30          3763467682  779340061263425537    60
#> 49236 2020-11-30          4707533116          2953955753     4
#> 49238 2020-11-30          4707533116           500964169     1
#> 49295 2020-11-30           500964169           112822128     1
#> 49296 2020-11-30           500964169 1215668032934117378     3
#> 49297 2020-11-30           500964169           143689068     3
#> 49298 2020-11-30           500964169          1579326655     1
#> 49299 2020-11-30           500964169           188998649     1
#> 49300 2020-11-30           500964169          2953955753     2
#> 49303 2020-11-30           500964169  779340061263425537     3
#> 49304 2020-11-30           500964169  879024609227665408     4
#> 49390 2020-11-30           593781032          2953955753     1
#> 49676 2020-11-30           950224302          2953955753     1
#> 49677 2020-11-30           950224302          3763467682     1
#> 49749 2020-12-31           110043455          2953955753     1
#> 49750 2020-12-31           110043455           500964169     1
#> 49759 2020-12-31           112822128          2953955753     1
#> 49763 2020-12-31           112822128           500964169     1
#> 49892 2020-12-31 1208861562410283010          2953955753     1
#> 49893 2020-12-31 1213976332750618624 1213976332750618624     5
#> 49894 2020-12-31 1213976332750618624           143689068     2
#> 49895 2020-12-31 1213976332750618624          2953955753    25
#> 49897 2020-12-31 1213976332750618624           500964169    25
#> 49903 2020-12-31 1215668032934117378          2953955753     1
#> 50020 2020-12-31           143689068          2953955753     4
#> 50022 2020-12-31           143689068           500964169     1
#> 50024 2020-12-31           143852870           339334958     2
#> 50142 2020-12-31          1579326655          2953955753     6
#> 50153 2020-12-31          1584076484          2953955753     2
#> 50155 2020-12-31          1584076484           500964169     3
#> 50156 2020-12-31          1584076484  779340061263425537     1
#> 50255 2020-12-31           172841079            37494271     1
#> 50263 2020-12-31           174801590            77009449     1
#> 50301 2020-12-31           185076776          2953955753    13
#> 50304 2020-12-31           185076776           500964169     3
#> 50305 2020-12-31           185076776  779340061263425537     1
#> 50315 2020-12-31           188998649          2953955753     4
#> 50316 2020-12-31           188998649           500964169     1
#> 50317 2020-12-31           188998649  879024609227665408     1
#> 50341 2020-12-31           200859218          1691697295     1
#> 50492 2020-12-31          2544467503          2953955753     1
#> 50494 2020-12-31          2544467503           935088068     1
#> 50615 2020-12-31           288342621           172841079     1
#> 50616 2020-12-31           288342621            17449646     1
#> 50617 2020-12-31           288342621           288342621     2
#> 50618 2020-12-31           288342621           313577888     1
#> 50619 2020-12-31           288342621            37494271     1
#> 50624 2020-12-31          2953955753 1215668032934117378     1
#> 50625 2020-12-31          2953955753           143689068     1
#> 50628 2020-12-31          2953955753           500964169    12
#> 50629 2020-12-31           297543171            37494271     4
#> 50630 2020-12-31           297543171            77009449     3
#> 50670 2020-12-31          3256203501           143852870     4
#> 50779 2020-12-31            34455479            37494271     2
#> 50780 2020-12-31            34455479            77009449    59
#> 50847 2020-12-31          3763467682 1208861562410283010     1
#> 50851 2020-12-31          3763467682           185076776     1
#> 50852 2020-12-31          3763467682          2953955753     7
#> 50853 2020-12-31          3763467682          3763467682     2
#> 50855 2020-12-31          3763467682           500964169    17
#> 50856 2020-12-31          3763467682  779340061263425537    55
#> 50857 2020-12-31          3763467682  879024609227665408     1
#> 50964 2020-12-31          4449919888           112822128     2
#> 50967 2020-12-31          4449919888           143689068     1
#> 50970 2020-12-31          4449919888          2841506345     1
#> 50971 2020-12-31          4449919888          2953955753     1
#> 51010 2020-12-31          4707533116           112822128     3
#> 51013 2020-12-31          4707533116           143689068     2
#> 51016 2020-12-31          4707533116          1579326655     1
#> 51018 2020-12-31          4707533116          2953955753     9
#> 51020 2020-12-31          4707533116           500964169     5
#> 51023 2020-12-31          4707533116  779340061263425537     1
#> 51071 2020-12-31           500964169 1215668032934117378     1
#> 51072 2020-12-31           500964169           143689068     1
#> 51073 2020-12-31           500964169          1579326655     1
#> 51075 2020-12-31           500964169           185076776     1
#> 51076 2020-12-31           500964169          2953955753     1
#> 51078 2020-12-31           500964169          4707533116     2
#> 51079 2020-12-31           500964169  779340061263425537     2
#> 51170 2020-12-31           593781032          2953955753     1
#> 51171 2020-12-31           593781032           500964169     4
#> 51295 2020-12-31            77009449            37494271     5
#> 51313 2020-12-31            84713190          4449919888     1
#> 51458 2020-12-31           950224302          2953955753     2
#> 51459 2020-12-31           950224302           500964169     1
get_network_data(category = c('province servants', 'others'),  start_date= "2020-10-01", end_date = "2020-12-31")
#> Error in quiet(download_list()) could not find function "quiet"