Data of every user in politicos en twitter shiny app (https://oderedes.shinyapps.io/politicosentwitter/)

A dataset containing information about every user downloaded

politicxs_data

Format

A dataset of 510 rows and 8 columns

user_id

Character variable with the unique ID of the Twitter user

screen_name

Character variable with the unique screen name of the Twitter user during may 2021

category

Character variable that shows the category of the user into the national politics during may 2021

description

Character variable that shows the description of the user in their account on Twitter during may 2021

name

Character variable that shows the name of the user in their account on Twitter during may 2021

gender

Character variable that shows the gender of the user in their account on Twitter

followers_count

Numeric variable that shows the count of followers during may 2021

profile_image_url

Character variable that shows the profile image of the user during may 2021

country

Character variable that shows the country where the user were elected or born

Source

Politicos en Twitter Shiny App (https://oderedes.shinyapps.io/politicosentwitter/)