Download the data of who's retweets who's

This function download all the retweets that every user on the list of politician on twitter made with each others

get_network_data(
  category = "all",
  start_date = "2021-01-01",
  end_date = "2021-03-31"
)

Arguments

category
A character vector with the category selected -"deputies","senators","national executive","others","province servants", "all"-

start_date
A character with the date where the retweets ocurr. This param read the date info in format 'yyyy-mm-dd'

end_date
A character with the date where the retweets stop. This param read the date info in format 'yyyy-mm-dd'

Examples

get_network_data(category = 'province servants',  start_date= "2020-10-01", end_date = "2020-12-31")
#>  Found 1000 records... Found 1020 records... Imported 1020 records. Simplifying into dataframe...
#>  Found 1000 records... Found 2000 records... Found 3000 records... Found 4000 records... Found 5000 records... Found 6000 records... Found 7000 records... Found 8000 records... Found 9000 records... Found 10000 records... Found 11000 records... Found 12000 records... Found 13000 records... Found 14000 records... Found 15000 records... Found 16000 records... Found 17000 records... Found 18000 records... Found 19000 records... Found 20000 records... Found 21000 records... Found 22000 records... Found 23000 records... Found 24000 records... Found 25000 records... Found 26000 records... Found 27000 records... Found 28000 records... Found 29000 records... Found 30000 records... Found 31000 records... Found 32000 records... Found 33000 records... Found 34000 records... Found 35000 records... Found 36000 records... Found 37000 records... Found 38000 records... Found 39000 records... Found 40000 records... Found 41000 records... Found 42000 records... Found 43000 records... Found 44000 records... Found 45000 records... Found 46000 records... Found 47000 records... Found 48000 records... Found 49000 records... Found 50000 records... Found 51000 records... Found 52000 records... Found 53000 records... Found 54000 records... Found 55000 records... Found 56000 records... Found 57000 records... Found 58000 records... Found 59000 records... Found 60000 records... Found 61000 records... Found 62000 records... Found 63000 records... Found 64000 records... Found 65000 records... Found 66000 records... Found 67000 records... Found 68000 records... Found 69000 records... Found 70000 records... Found 71000 records... Found 72000 records... Found 73000 records... Found 74000 records... Found 75000 records... Found 76000 records... Found 77000 records... Found 78000 records... Found 79000 records... Found 80000 records... Found 81000 records... Found 82000 records... Found 83000 records... Found 84000 records... Found 85000 records... Found 85865 records... Imported 85865 records. Simplifying into dataframe...
#>       month_year             user_id     retweet_user_id value
#> 45911 2020-10-31           110043455           500964169     1
#> 46101 2020-10-31 1213976332750618624          2953955753     5
#> 46105 2020-10-31 1213976332750618624           500964169     1
#> 46106 2020-10-31 1213976332750618624  779340061263425537     1
#> 46288 2020-10-31          1475504383          2953955753     2
#> 46290 2020-10-31          1475504383           500964169     1
#> 46403 2020-10-31          1579326655           143689068     1
#> 46407 2020-10-31          1579326655          2953955753     3
#> 46416 2020-10-31          1584076484          2953955753     2
#> 46419 2020-10-31          1584076484           500964169     3
#> 46591 2020-10-31           174801590 1239520164699398145     1
#> 46627 2020-10-31           185076776          2953955753    45
#> 46630 2020-10-31           185076776           500964169     5
#> 46632 2020-10-31           185076776  879024609227665408     1
#> 46649 2020-10-31           188998649          2953955753     7
#> 46650 2020-10-31           188998649          4707533116     1
#> 46690 2020-10-31           200859218          1691697295     2
#> 47016 2020-10-31           288342621           172841079     1
#> 47017 2020-10-31           288342621            37494271     2
#> 47029 2020-10-31          2953955753           143689068     1
#> 47034 2020-10-31          2953955753           500964169     7
#> 47036 2020-10-31          2953955753  879024609227665408     2
#> 47037 2020-10-31           297543171            37494271     5
#> 47038 2020-10-31           297543171            77009449     2
#> 47234 2020-10-31            34455479            37494271     3
#> 47235 2020-10-31            34455479            77009449    54
#> 47320 2020-10-31          3763467682           143689068     3
#> 47323 2020-10-31          3763467682          1579326655     3
#> 47324 2020-10-31          3763467682          1584076484     4
#> 47325 2020-10-31          3763467682           185076776     1
#> 47326 2020-10-31          3763467682          2953955753    17
#> 47329 2020-10-31          3763467682           500964169     4
#> 47333 2020-10-31          3763467682  779340061263425537    74
#> 47492 2020-10-31          4707533116           143689068     3
#> 47493 2020-10-31          4707533116          1475504383     1
#> 47496 2020-10-31          4707533116          2953955753     3
#> 47497 2020-10-31          4707533116           500964169     2
#> 47556 2020-10-31           500964169 1215668032934117378     1
#> 47557 2020-10-31           500964169          1579326655     2
#> 47558 2020-10-31           500964169           188998649     1
#> 47559 2020-10-31           500964169          2953955753     1
#> 47562 2020-10-31           500964169  779340061263425537     1
#> 47563 2020-10-31           500964169  879024609227665408     2
#> 48051 2020-10-31           950224302           185076776     1
#> 48052 2020-10-31           950224302          2953955753     4
#> 48259 2020-11-30 1213976332750618624          2953955753     6
#> 48263 2020-11-30 1213976332750618624           500964169     6
#> 48482 2020-11-30          1579326655           143689068     1
#> 48485 2020-11-30          1579326655          2953955753     2
#> 48493 2020-11-30          1584076484           500964169     4
#> 48494 2020-11-30          1584076484  779340061263425537     2
#> 48495 2020-11-30          1584076484           950224302     1
#> 48632 2020-11-30           185076776          2953955753    17
#> 48633 2020-11-30           185076776           500964169     2
#> 48634 2020-11-30           185076776           593781032     1
#> 48636 2020-11-30           188998649          2953955753     4
#> 48655 2020-11-30           200859218          1691697295     1
#> 48786 2020-11-30          2544467503           143852870     1
#> 48897 2020-11-30           288342621           164368516     1
#> 48898 2020-11-30           288342621           172841079     1
#> 48899 2020-11-30           288342621            17449646     1
#> 48900 2020-11-30           288342621           297543171     1
#> 48901 2020-11-30           288342621            37494271     2
#> 48903 2020-11-30           288342621            77009449     2
#> 48910 2020-11-30          2953955753          1579326655     1
#> 48912 2020-11-30          2953955753           500964169     7
#> 48915 2020-11-30          2953955753  879024609227665408     1
#> 48916 2020-11-30           297543171            37494271     5
#> 48917 2020-11-30           297543171            77009449     2
#> 48963 2020-11-30          3256203501          2544467503     2
#> 49032 2020-11-30            34455479            37494271     2
#> 49033 2020-11-30            34455479            77009449    63
#> 49073 2020-11-30          3763467682          1584076484     1
#> 49075 2020-11-30          3763467682          2589783002     1
#> 49076 2020-11-30          3763467682          2953955753     3
#> 49079 2020-11-30          3763467682           500964169     4
#> 49082 2020-11-30          3763467682  779340061263425537    60
#> 49236 2020-11-30          4707533116          2953955753     4
#> 49238 2020-11-30          4707533116           500964169     1
#> 49295 2020-11-30           500964169           112822128     1
#> 49296 2020-11-30           500964169 1215668032934117378     3
#> 49297 2020-11-30           500964169           143689068     3
#> 49298 2020-11-30           500964169          1579326655     1
#> 49299 2020-11-30           500964169           188998649     1
#> 49300 2020-11-30           500964169          2953955753     2
#> 49303 2020-11-30           500964169  779340061263425537     3
#> 49304 2020-11-30           500964169  879024609227665408     4
#> 49390 2020-11-30           593781032          2953955753     1
#> 49676 2020-11-30           950224302          2953955753     1
#> 49677 2020-11-30           950224302          3763467682     1
#> 49749 2020-12-31           110043455          2953955753     1
#> 49750 2020-12-31           110043455           500964169     1
#> 49759 2020-12-31           112822128          2953955753     1
#> 49763 2020-12-31           112822128           500964169     1
#> 49892 2020-12-31 1208861562410283010          2953955753     1
#> 49893 2020-12-31 1213976332750618624 1213976332750618624     5
#> 49894 2020-12-31 1213976332750618624           143689068     2
#> 49895 2020-12-31 1213976332750618624          2953955753    25
#> 49897 2020-12-31 1213976332750618624           500964169    25
#> 49903 2020-12-31 1215668032934117378          2953955753     1
#> 50020 2020-12-31           143689068          2953955753     4
#> 50022 2020-12-31           143689068           500964169     1
#> 50024 2020-12-31           143852870           339334958     2
#> 50142 2020-12-31          1579326655          2953955753     6
#> 50153 2020-12-31          1584076484          2953955753     2
#> 50155 2020-12-31          1584076484           500964169     3
#> 50156 2020-12-31          1584076484  779340061263425537     1
#> 50255 2020-12-31           172841079            37494271     1
#> 50263 2020-12-31           174801590            77009449     1
#> 50301 2020-12-31           185076776          2953955753    13
#> 50304 2020-12-31           185076776           500964169     3
#> 50305 2020-12-31           185076776  779340061263425537     1
#> 50315 2020-12-31           188998649          2953955753     4
#> 50316 2020-12-31           188998649           500964169     1
#> 50317 2020-12-31           188998649  879024609227665408     1
#> 50341 2020-12-31           200859218          1691697295     1
#> 50492 2020-12-31          2544467503          2953955753     1
#> 50494 2020-12-31          2544467503           935088068     1
#> 50615 2020-12-31           288342621           172841079     1
#> 50616 2020-12-31           288342621            17449646     1
#> 50617 2020-12-31           288342621           288342621     2
#> 50618 2020-12-31           288342621           313577888     1
#> 50619 2020-12-31           288342621            37494271     1
#> 50624 2020-12-31          2953955753 1215668032934117378     1
#> 50625 2020-12-31          2953955753           143689068     1
#> 50628 2020-12-31          2953955753           500964169    12
#> 50629 2020-12-31           297543171            37494271     4
#> 50630 2020-12-31           297543171            77009449     3
#> 50670 2020-12-31          3256203501           143852870     4
#> 50779 2020-12-31            34455479            37494271     2
#> 50780 2020-12-31            34455479            77009449    59
#> 50847 2020-12-31          3763467682 1208861562410283010     1
#> 50851 2020-12-31          3763467682           185076776     1
#> 50852 2020-12-31          3763467682          2953955753     7
#> 50853 2020-12-31          3763467682          3763467682     2
#> 50855 2020-12-31          3763467682           500964169    17
#> 50856 2020-12-31          3763467682  779340061263425537    55
#> 50857 2020-12-31          3763467682  879024609227665408     1
#> 50964 2020-12-31          4449919888           112822128     2
#> 50967 2020-12-31          4449919888           143689068     1
#> 50970 2020-12-31          4449919888          2841506345     1
#> 50971 2020-12-31          4449919888          2953955753     1
#> 51010 2020-12-31          4707533116           112822128     3
#> 51013 2020-12-31          4707533116           143689068     2
#> 51016 2020-12-31          4707533116          1579326655     1
#> 51018 2020-12-31          4707533116          2953955753     9
#> 51020 2020-12-31          4707533116           500964169     5
#> 51023 2020-12-31          4707533116  779340061263425537     1
#> 51071 2020-12-31           500964169 1215668032934117378     1
#> 51072 2020-12-31           500964169           143689068     1
#> 51073 2020-12-31           500964169          1579326655     1
#> 51075 2020-12-31           500964169           185076776     1
#> 51076 2020-12-31           500964169          2953955753     1
#> 51078 2020-12-31           500964169          4707533116     2
#> 51079 2020-12-31           500964169  779340061263425537     2
#> 51170 2020-12-31           593781032          2953955753     1
#> 51171 2020-12-31           593781032           500964169     4
#> 51295 2020-12-31            77009449            37494271     5
#> 51313 2020-12-31            84713190          4449919888     1
#> 51458 2020-12-31           950224302          2953955753     2
#> 51459 2020-12-31           950224302           500964169     1
get_network_data(category = c('province servants', 'others'),  start_date= "2020-10-01", end_date = "2020-12-31")
#>  Found 1000 records... Found 1020 records... Imported 1020 records. Simplifying into dataframe...
#>  Found 1000 records... Found 2000 records... Found 3000 records... Found 4000 records... Found 5000 records... Found 6000 records... Found 7000 records... Found 8000 records... Found 9000 records... Found 10000 records... Found 11000 records... Found 12000 records... Found 13000 records... Found 14000 records... Found 15000 records... Found 16000 records... Found 17000 records... Found 18000 records... Found 19000 records... Found 20000 records... Found 21000 records... Found 22000 records... Found 23000 records... Found 24000 records... Found 25000 records... Found 26000 records... Found 27000 records... Found 28000 records... Found 29000 records... Found 30000 records... Found 31000 records... Found 32000 records... Found 33000 records... Found 34000 records... Found 35000 records... Found 36000 records... Found 37000 records... Found 38000 records... Found 39000 records... Found 40000 records... Found 41000 records... Found 42000 records... Found 43000 records... Found 44000 records... Found 45000 records... Found 46000 records... Found 47000 records... Found 48000 records... Found 49000 records... Found 50000 records... Found 51000 records... Found 52000 records... Found 53000 records... Found 54000 records... Found 55000 records... Found 56000 records... Found 57000 records... Found 58000 records... Found 59000 records... Found 60000 records... Found 61000 records... Found 62000 records... Found 63000 records... Found 64000 records... Found 65000 records... Found 66000 records... Found 67000 records... Found 68000 records... Found 69000 records... Found 70000 records... Found 71000 records... Found 72000 records... Found 73000 records... Found 74000 records... Found 75000 records... Found 76000 records... Found 77000 records... Found 78000 records... Found 79000 records... Found 80000 records... Found 81000 records... Found 82000 records... Found 83000 records... Found 84000 records... Found 85000 records... Found 85865 records... Imported 85865 records. Simplifying into dataframe...
#>       month_year             user_id     retweet_user_id value
#> 45826 2020-10-31           101884357           101884357     1
#> 45853 2020-10-31           105180935            18143118     1
#> 45854 2020-10-31           105180935            24900072     1
#> 45873 2020-10-31 1082401224312406018          1383349418     1
#> 45883 2020-10-31 1082401224312406018          2953955753     2
#> 45890 2020-10-31 1082401224312406018            61631756     2
#> 45907 2020-10-31 1095773453855137794 1095773453855137794    20
#> 45908 2020-10-31 1095773453855137794           175273213     1
#> 45911 2020-10-31           110043455           500964169     1
#> 45919 2020-10-31 1132308934948282368 1082401224312406018     6
#> 45920 2020-10-31 1132308934948282368           112822128     5
#> 45926 2020-10-31 1132308934948282368          1383349418     1
#> 45928 2020-10-31 1132308934948282368           143689068     7
#> 45929 2020-10-31 1132308934948282368           143852870     1
#> 45930 2020-10-31 1132308934948282368           145685744     1
#> 45934 2020-10-31 1132308934948282368          1579326655     2
#> 45949 2020-10-31 1132308934948282368          2544467503     1
#> 45955 2020-10-31 1132308934948282368          2953955753    36
#> 45957 2020-10-31 1132308934948282368          3028492511     1
#> 45958 2020-10-31 1132308934948282368          3256203501     1
#> 45968 2020-10-31 1132308934948282368           500964169     1
#> 45971 2020-10-31 1132308934948282368           586933252     1
#> 45976 2020-10-31 1132308934948282368  779498837236916224     1
#> 45987 2020-10-31           114582574           114582574     3
#> 46008 2020-10-31           114582574           339334958     7
#> 46016 2020-10-31           114582574           607590053     5
#> 46018 2020-10-31           114582574  931564592328781824     3
#> 46019 2020-10-31           114582574            93763185     3
#> 46042 2020-10-31           116313517 1132308934948282368     1
#> 46101 2020-10-31 1213976332750618624          2953955753     5
#> 46105 2020-10-31 1213976332750618624           500964169     1
#> 46106 2020-10-31 1213976332750618624  779340061263425537     1
#> 46168 2020-10-31 1268302274801131520            18143118     1
#> 46188 2020-10-31           130574455            40858089     1
#> 46251 2020-10-31           139266780           152835465     1
#> 46252 2020-10-31           139266780           259681081     1
#> 46253 2020-10-31           139266780           730926774     1
#> 46285 2020-10-31          1469894064            24900072     1
#> 46286 2020-10-31          1469894064           607590053     1
#> 46288 2020-10-31          1475504383          2953955753     2
#> 46290 2020-10-31          1475504383           500964169     1
#> 46307 2020-10-31           148752676           116313517     1
#> 46308 2020-10-31           148752676           148752676    29
#> 46336 2020-10-31           150123282           105180935     1
#> 46394 2020-10-31           157479149           299298402     1
#> 46397 2020-10-31          1578961946           174801590     1
#> 46400 2020-10-31          1578961946           485988630     2
#> 46401 2020-10-31          1579326655 1132308934948282368     2
#> 46403 2020-10-31          1579326655           143689068     1
#> 46407 2020-10-31          1579326655          2953955753     3
#> 46414 2020-10-31          1584076484 1132308934948282368     3
#> 46416 2020-10-31          1584076484          2953955753     2
#> 46419 2020-10-31          1584076484           500964169     3
#> 46456 2020-10-31           162099154 1132308934948282368     4
#> 46458 2020-10-31           162099154          1383349418     1
#> 46460 2020-10-31           162099154           143852870     1
#> 46461 2020-10-31           162099154           145685744     1
#> 46464 2020-10-31           162099154           162099154     2
#> 46465 2020-10-31           162099154          1691697295     1
#> 46466 2020-10-31           162099154           187354449     3
#> 46470 2020-10-31           162099154           208612518     1
#> 46473 2020-10-31           162099154          2544467503     1
#> 46474 2020-10-31           162099154          2589783002     1
#> 46476 2020-10-31           162099154          2953955753     2
#> 46477 2020-10-31           162099154           297784596     1
#> 46478 2020-10-31           162099154          3028492511     1
#> 46479 2020-10-31           162099154          3256203501     1
#> 46480 2020-10-31           162099154          3342878295     1
#> 46485 2020-10-31           162099154          4179135675     1
#> 46488 2020-10-31           162099154          4449919888     2
#> 46490 2020-10-31           162099154          4707533116     1
#> 46492 2020-10-31           162099154           586933252     1
#> 46493 2020-10-31           162099154            61631756     1
#> 46495 2020-10-31           162099154  779498837236916224     1
#> 46591 2020-10-31           174801590 1239520164699398145     1
#> 46593 2020-10-31           174801590          1578961946     2
#> 46612 2020-10-31            18143118           114582574     1
#> 46613 2020-10-31            18143118           124222509     1
#> 46616 2020-10-31            18143118            24900072     4
#> 46617 2020-10-31            18143118            54414081     1
#> 46618 2020-10-31            18143118            77009449     1
#> 46627 2020-10-31           185076776          2953955753    45
#> 46630 2020-10-31           185076776           500964169     5
#> 46632 2020-10-31           185076776  879024609227665408     1
#> 46649 2020-10-31           188998649          2953955753     7
#> 46650 2020-10-31           188998649          4707533116     1
#> 46690 2020-10-31           200859218          1691697295     2
#> 46739 2020-10-31          2192397067           299298402     6
#> 46741 2020-10-31          2192397067            84713190     1
#> 46804 2020-10-31            23677018            54414081     3
#> 46891 2020-10-31           257432060           175422899     6
#> 46892 2020-10-31           257432060           257432060    51
#> 46894 2020-10-31           257432060          3167567897    15
#> 46895 2020-10-31           257432060            65360466     7
#> 47016 2020-10-31           288342621           172841079     1
#> 47017 2020-10-31           288342621            37494271     2
#> 47028 2020-10-31          2953955753 1132308934948282368     5
#> 47029 2020-10-31          2953955753           143689068     1
#> 47034 2020-10-31          2953955753           500964169     7
#> 47036 2020-10-31          2953955753  879024609227665408     2
#> 47037 2020-10-31           297543171            37494271     5
#> 47038 2020-10-31           297543171            77009449     2
#> 47039 2020-10-31           297784596           297784596     1
#> 47043 2020-10-31           299298402          2192397067     4
#> 47045 2020-10-31           299298402           299298402    10
#> 47051 2020-10-31           299298402            84713190     2
#> 47058 2020-10-31           301464513           257432060     1
#> 47060 2020-10-31           301464513           332993154    23
#> 47136 2020-10-31           330546547          1578961946     1
#> 47137 2020-10-31           330546547            24900072     2
#> 47138 2020-10-31           330546547            54414081     2
#> 47139 2020-10-31          3313610230 1132308934948282368     1
#> 47141 2020-10-31          3313610230           148752676     1
#> 47151 2020-10-31          3313610230          2953955753     1
#> 47152 2020-10-31          3313610230           297784596     1
#> 47158 2020-10-31          3313610230          4449919888     4
#> 47160 2020-10-31          3313610230           535633686     1
#> 47161 2020-10-31          3313610230  779498837236916224     1
#> 47234 2020-10-31            34455479            37494271     3
#> 47235 2020-10-31            34455479            77009449    54
#> 47239 2020-10-31           349968963            34455479     1
#> 47240 2020-10-31           349968963            84713190     1
#> 47241 2020-10-31           351953986 1074441014625124352     3
#> 47243 2020-10-31           351953986           351953986    80
#> 47244 2020-10-31           351953986          4020276615     2
#> 47318 2020-10-31          3763467682 1132308934948282368     2
#> 47320 2020-10-31          3763467682           143689068     3
#> 47323 2020-10-31          3763467682          1579326655     3
#> 47324 2020-10-31          3763467682          1584076484     4
#> 47325 2020-10-31          3763467682           185076776     1
#> 47326 2020-10-31          3763467682          2953955753    17
#> 47329 2020-10-31          3763467682           500964169     4
#> 47333 2020-10-31          3763467682  779340061263425537    74
#> 47355 2020-10-31           398221268            38494089     1
#> 47365 2020-10-31            40858089 1043198260092039169     1
#> 47369 2020-10-31            40858089            18143118     2
#> 47371 2020-10-31            40858089           217169311     1
#> 47372 2020-10-31            40858089            24900072     3
#> 47373 2020-10-31            40858089           257460186     1
#> 47374 2020-10-31            40858089           330546547     1
#> 47375 2020-10-31            40858089            38494089     2
#> 47376 2020-10-31            40858089            77009449     5
#> 47382 2020-10-31           416721882            24900072     3
#> 47384 2020-10-31          4198001913           284783298     1
#> 47492 2020-10-31          4707533116           143689068     3
#> 47493 2020-10-31          4707533116          1475504383     1
#> 47496 2020-10-31          4707533116          2953955753     3
#> 47497 2020-10-31          4707533116           500964169     2
#> 47498 2020-10-31          4707533116  724991849912500224     1
#> 47536 2020-10-31           485988630          1578961946     9
#> 47537 2020-10-31           485988630           174801590     3
#> 47544 2020-10-31           485988630           485988630     3
#> 47556 2020-10-31           500964169 1215668032934117378     1
#> 47557 2020-10-31           500964169          1579326655     2
#> 47558 2020-10-31           500964169           188998649     1
#> 47559 2020-10-31           500964169          2953955753     1
#> 47562 2020-10-31           500964169  779340061263425537     1
#> 47563 2020-10-31           500964169  879024609227665408     2
#> 47602 2020-10-31            54414081           151953199     1
#> 47603 2020-10-31            54414081           330546547     1
#> 47605 2020-10-31            54414081           416721882     1
#> 47669 2020-10-31            60534269            38494089     1
#> 47676 2020-10-31            63511700 1132308934948282368     2
#> 47681 2020-10-31            65360466           175422899    33
#> 47682 2020-10-31            65360466           257432060    28
#> 47684 2020-10-31            65360466          3167567897     4
#> 47815 2020-10-31            77897441           576528582     1
#> 47852 2020-10-31             8565932 1082401224312406018     1
#> 47853 2020-10-31             8565932 1132308934948282368    17
#> 47865 2020-10-31             8565932          2953955753    13
#> 47871 2020-10-31             8565932  724991849912500224     3
#> 48051 2020-10-31           950224302           185076776     1
#> 48052 2020-10-31           950224302          2953955753     4
#> 48053 2020-10-31            95970921            18143118     1
#> 48054 2020-10-31            95970921  931564592328781824     1
#> 48099 2020-11-30           105180935            24900072     2
#> 48100 2020-11-30           105180935            54414081     1
#> 48119 2020-11-30 1082401224312406018           586933252     2
#> 48128 2020-11-30 1095773453855137794 1095773453855137794    14
#> 48132 2020-11-30 1132308934948282368 1082401224312406018     3
#> 48133 2020-11-30 1132308934948282368           112822128     1
#> 48140 2020-11-30 1132308934948282368           143689068     1
#> 48141 2020-11-30 1132308934948282368           143852870     1
#> 48144 2020-11-30 1132308934948282368          1579326655     1
#> 48167 2020-11-30 1132308934948282368          2953955753    38
#> 48178 2020-11-30 1132308934948282368           500964169     4
#> 48197 2020-11-30           114582574            18143118     2
#> 48209 2020-11-30           114582574           339334958     7
#> 48217 2020-11-30           114582574           607590053     5
#> 48220 2020-11-30           114582574  931564592328781824     6
#> 48221 2020-11-30           114582574            93763185     6
#> 48259 2020-11-30 1213976332750618624          2953955753     6
#> 48263 2020-11-30 1213976332750618624           500964169     6
#> 48298 2020-11-30 1268302274801131520           151953199     1
#> 48299 2020-11-30 1268302274801131520            54414081     1
#> 48306 2020-11-30           130574455            54414081     1
#> 48359 2020-11-30           139266780           174010839     1
#> 48385 2020-11-30          1469894064            24900072     2
#> 48386 2020-11-30          1469894064           576528582     1
#> 48396 2020-11-30           148752676           148752676    24
#> 48401 2020-11-30           148752676           527663666     1
#> 48416 2020-11-30           150123282           576528582     1
#> 48438 2020-11-30           154731462           154731462     1
#> 48468 2020-11-30           157479149           157479149     4
#> 48469 2020-11-30           157479149           299298402     1
#> 48472 2020-11-30          1578961946           174801590     1
#> 48477 2020-11-30          1578961946           485988630     1
#> 48480 2020-11-30          1579326655 1132308934948282368     2
#> 48482 2020-11-30          1579326655           143689068     1
#> 48485 2020-11-30          1579326655          2953955753     2
#> 48493 2020-11-30          1584076484           500964169     4
#> 48494 2020-11-30          1584076484  779340061263425537     2
#> 48495 2020-11-30          1584076484           950224302     1
#> 48517 2020-11-30           162099154           134232117     2
#> 48524 2020-11-30           162099154           297784596     1
#> 48529 2020-11-30           162099154          4179135675     1
#> 48532 2020-11-30           162099154            61631756     1
#> 48566 2020-11-30           164774579          4198001913     1
#> 48581 2020-11-30           174010839           288342621     1
#> 48593 2020-11-30           174801590          1578961946     2
#> 48594 2020-11-30           174801590           485988630     1
#> 48613 2020-11-30            18143118           105180935     1
#> 48615 2020-11-30            18143118            24900072     2
#> 48618 2020-11-30            18143118            54414081     2
#> 48619 2020-11-30            18143118            77009449     4
#> 48632 2020-11-30           185076776          2953955753    17
#> 48633 2020-11-30           185076776           500964169     2
#> 48634 2020-11-30           185076776           593781032     1
#> 48636 2020-11-30           188998649          2953955753     4
#> 48655 2020-11-30           200859218          1691697295     1
#> 48687 2020-11-30          2192397067           143689068     1
#> 48689 2020-11-30          2192397067           299298402     1
#> 48725 2020-11-30            23677018           105180935     3
#> 48726 2020-11-30            23677018           150123282     1
#> 48728 2020-11-30            23677018            24900072     1
#> 48729 2020-11-30            23677018            54414081     3
#> 48760 2020-11-30            24900072          1469894064     1
#> 48786 2020-11-30          2544467503           143852870     1
#> 48789 2020-11-30           257432060           175422899     6
#> 48790 2020-11-30           257432060           257432060    55
#> 48792 2020-11-30           257432060          3167567897     5
#> 48793 2020-11-30           257432060           332993154     1
#> 48794 2020-11-30           257432060            65360466     8
#> 48897 2020-11-30           288342621           164368516     1
#> 48898 2020-11-30           288342621           172841079     1
#> 48899 2020-11-30           288342621            17449646     1
#> 48900 2020-11-30           288342621           297543171     1
#> 48901 2020-11-30           288342621            37494271     2
#> 48902 2020-11-30           288342621            54414081     1
#> 48903 2020-11-30           288342621            77009449     2
#> 48910 2020-11-30          2953955753          1579326655     1
#> 48912 2020-11-30          2953955753           500964169     7
#> 48915 2020-11-30          2953955753  879024609227665408     1
#> 48916 2020-11-30           297543171            37494271     5
#> 48917 2020-11-30           297543171            77009449     2
#> 48918 2020-11-30           297784596           297784596     1
#> 48921 2020-11-30           299298402 1215668032934117378     1
#> 48923 2020-11-30           299298402          2192397067     1
#> 48924 2020-11-30           299298402           299298402     5
#> 48926 2020-11-30           301464513           332993154     8
#> 48963 2020-11-30          3256203501          2544467503     2
#> 48979 2020-11-30          3313610230           101884357     1
#> 48988 2020-11-30          3313610230           527663666     1
#> 49032 2020-11-30            34455479            37494271     2
#> 49033 2020-11-30            34455479            77009449    63
#> 49035 2020-11-30           351953986 1074441014625124352     3
#> 49036 2020-11-30           351953986           351953986    69
#> 49037 2020-11-30           351953986          4020276615     2
#> 49068 2020-11-30          3763467682 1132308934948282368     4
#> 49073 2020-11-30          3763467682          1584076484     1
#> 49075 2020-11-30          3763467682          2589783002     1
#> 49076 2020-11-30          3763467682          2953955753     3
#> 49079 2020-11-30          3763467682           500964169     4
#> 49081 2020-11-30          3763467682  724991849912500224     1
#> 49082 2020-11-30          3763467682  779340061263425537    60
#> 49102 2020-11-30          4020276615 1074441014625124352     1
#> 49103 2020-11-30          4020276615 1095773453855137794     3
#> 49104 2020-11-30          4020276615           155434166     1
#> 49105 2020-11-30          4020276615           175273213     8
#> 49106 2020-11-30          4020276615           351953986    19
#> 49107 2020-11-30          4020276615          4020276615    44
#> 49108 2020-11-30          4020276615            56818132     7
#> 49118 2020-11-30            40858089 1043198260092039169     1
#> 49121 2020-11-30            40858089            18143118     1
#> 49123 2020-11-30            40858089           288342621     1
#> 49124 2020-11-30            40858089           299371033     1
#> 49125 2020-11-30            40858089           330546547     2
#> 49130 2020-11-30          4198001913           284783298     2
#> 49231 2020-11-30          4707533116 1132308934948282368     1
#> 49236 2020-11-30          4707533116          2953955753     4
#> 49238 2020-11-30          4707533116           500964169     1
#> 49271 2020-11-30           485988630          1578961946    22
#> 49272 2020-11-30           485988630           174801590     1
#> 49275 2020-11-30           485988630           249976489     3
#> 49281 2020-11-30           485988630           485988630     2
#> 49295 2020-11-30           500964169           112822128     1
#> 49296 2020-11-30           500964169 1215668032934117378     3
#> 49297 2020-11-30           500964169           143689068     3
#> 49298 2020-11-30           500964169          1579326655     1
#> 49299 2020-11-30           500964169           188998649     1
#> 49300 2020-11-30           500964169          2953955753     2
#> 49303 2020-11-30           500964169  779340061263425537     3
#> 49304 2020-11-30           500964169  879024609227665408     4
#> 49341 2020-11-30            54414081            18143118     1
#> 49390 2020-11-30           593781032          2953955753     1
#> 49400 2020-11-30            60534269           124222509     1
#> 49408 2020-11-30            63511700 1132308934948282368     3
#> 49412 2020-11-30            65360466           175422899    21
#> 49413 2020-11-30            65360466           257432060    33
#> 49415 2020-11-30            65360466          3167567897     7
#> 49416 2020-11-30            65360466           332993154     1
#> 49511 2020-11-30           789856765           191209101     2
#> 49529 2020-11-30            84713190 1132308934948282368     1
#> 49533 2020-11-30            84713190          2192397067     1
#> 49534 2020-11-30            84713190           299298402     1
#> 49536 2020-11-30             8565932           112822128     1
#> 49537 2020-11-30             8565932 1132308934948282368     5
#> 49546 2020-11-30             8565932          2953955753    11
#> 49550 2020-11-30             8565932          4707533116     1
#> 49551 2020-11-30             8565932           500964169     1
#> 49553 2020-11-30             8565932  724991849912500224     1
#> 49676 2020-11-30           950224302          2953955753     1
#> 49677 2020-11-30           950224302          3763467682     1
#> 49680 2020-11-30            95970921            24900072     1
#> 49681 2020-11-30            95970921            54414081     3
#> 49682 2020-11-30            95970921  931564592328781824     1
#> 49706 2020-12-31           101884357           101884357     1
#> 49711 2020-12-31 1043198260092039169            54414081     1
#> 49721 2020-12-31           105180935 1268302274801131520     1
#> 49723 2020-12-31           105180935            18143118     1
#> 49724 2020-12-31           105180935            24900072     1
#> 49745 2020-12-31 1095773453855137794 1095773453855137794    13
#> 49746 2020-12-31 1095773453855137794           143502765     1
#> 49747 2020-12-31 1095773453855137794           175273213     1
#> 49748 2020-12-31 1095773453855137794          4020276615     6
#> 49749 2020-12-31           110043455          2953955753     1
#> 49750 2020-12-31           110043455           500964169     1
#> 49756 2020-12-31           112822128 1132308934948282368     1
#> 49759 2020-12-31           112822128          2953955753     1
#> 49763 2020-12-31           112822128           500964169     1
#> 49765 2020-12-31           112822128  724991849912500224     1
#> 49766 2020-12-31 1132308934948282368           112822128     1
#> 49772 2020-12-31 1132308934948282368           134232117     1
#> 49774 2020-12-31 1132308934948282368          1383349418     1
#> 49776 2020-12-31 1132308934948282368           143689068     2
#> 49777 2020-12-31 1132308934948282368           143852870     2
#> 49781 2020-12-31 1132308934948282368          1579326655     1
#> 49785 2020-12-31 1132308934948282368           208612518     1
#> 49792 2020-12-31 1132308934948282368          2544467503     2
#> 49797 2020-12-31 1132308934948282368          2953955753    36
#> 49798 2020-12-31 1132308934948282368          3028492511     2
#> 49799 2020-12-31 1132308934948282368          3256203501     2
#> 49807 2020-12-31 1132308934948282368          4179135675     1
#> 49811 2020-12-31 1132308934948282368           500964169     5
#> 49824 2020-12-31           114582574           114582574     1
#> 49839 2020-12-31           114582574           339246511     1
#> 49840 2020-12-31           114582574           339334958     8
#> 49847 2020-12-31           114582574           607590053     8
#> 49850 2020-12-31           114582574  931564592328781824     6
#> 49851 2020-12-31           114582574            93763185     2
#> 49892 2020-12-31 1208861562410283010          2953955753     1
#> 49893 2020-12-31 1213976332750618624 1213976332750618624     5
#> 49894 2020-12-31 1213976332750618624           143689068     2
#> 49895 2020-12-31 1213976332750618624          2953955753    25
#> 49897 2020-12-31 1213976332750618624           500964169    25
#> 49903 2020-12-31 1215668032934117378          2953955753     1
#> 50015 2020-12-31           143689068 1132308934948282368     3
#> 50020 2020-12-31           143689068          2953955753     4
#> 50022 2020-12-31           143689068           500964169     1
#> 50024 2020-12-31           143852870           339334958     2
#> 50052 2020-12-31          1469894064          1469894064     1
#> 50072 2020-12-31           148752676           148752676     9
#> 50073 2020-12-31           148752676           191209101     1
#> 50084 2020-12-31           150123282            24900072     1
#> 50086 2020-12-31           150123282            54414081     1
#> 50128 2020-12-31           157479149           157479149     1
#> 50129 2020-12-31           157479149           299298402     2
#> 50130 2020-12-31           157479149           525264466     2
#> 50134 2020-12-31          1579326655 1132308934948282368     2
#> 50142 2020-12-31          1579326655          2953955753     6
#> 50149 2020-12-31          1584076484 1132308934948282368     1
#> 50153 2020-12-31          1584076484          2953955753     2
#> 50155 2020-12-31          1584076484           500964169     3
#> 50156 2020-12-31          1584076484  779340061263425537     1
#> 50170 2020-12-31           162099154           134232117     1
#> 50172 2020-12-31           162099154          1383349418     5
#> 50174 2020-12-31           162099154           143852870     1
#> 50176 2020-12-31           162099154           162099154     2
#> 50177 2020-12-31           162099154          1691697295     1
#> 50181 2020-12-31           162099154          2589783002     1
#> 50186 2020-12-31           162099154          2953955753     1
#> 50187 2020-12-31           162099154          3028492511     2
#> 50188 2020-12-31           162099154          3256203501     1
#> 50190 2020-12-31           162099154          3342878295     1
#> 50195 2020-12-31           162099154           586933252     1
#> 50196 2020-12-31           162099154           607590053     1
#> 50197 2020-12-31           162099154  779498837236916224     1
#> 50198 2020-12-31           162099154           935088068     1
#> 50255 2020-12-31           172841079            37494271     1
#> 50263 2020-12-31           174801590            77009449     1
#> 50264 2020-12-31           175422899           257432060     1
#> 50282 2020-12-31            18143118 1239520164699398145     1
#> 50285 2020-12-31            18143118            24900072     2
#> 50286 2020-12-31            18143118           330546547     1
#> 50288 2020-12-31            18143118            48568308     2
#> 50289 2020-12-31            18143118            54414081     2
#> 50290 2020-12-31            18143118           576528582     1
#> 50291 2020-12-31            18143118            60534269     1
#> 50293 2020-12-31            18143118            77009449     2
#> 50301 2020-12-31           185076776          2953955753    13
#> 50304 2020-12-31           185076776           500964169     3
#> 50305 2020-12-31           185076776  779340061263425537     1
#> 50315 2020-12-31           188998649          2953955753     4
#> 50316 2020-12-31           188998649           500964169     1
#> 50317 2020-12-31           188998649  879024609227665408     1
#> 50341 2020-12-31           200859218          1691697295     1
#> 50372 2020-12-31          2192397067          2192397067     1
#> 50375 2020-12-31          2192397067           299298402     2
#> 50429 2020-12-31            23677018           105180935     1
#> 50465 2020-12-31            24900072           416721882     1
#> 50492 2020-12-31          2544467503          2953955753     1
#> 50494 2020-12-31          2544467503           935088068     1
#> 50495 2020-12-31           257432060           175422899     3
#> 50496 2020-12-31           257432060           257432060    31
#> 50498 2020-12-31           257432060          3167567897    10
#> 50499 2020-12-31           257432060            65360466     2
#> 50615 2020-12-31           288342621           172841079     1
#> 50616 2020-12-31           288342621            17449646     1
#> 50617 2020-12-31           288342621           288342621     2
#> 50618 2020-12-31           288342621           313577888     1
#> 50619 2020-12-31           288342621            37494271     1
#> 50623 2020-12-31          2953955753 1132308934948282368     1
#> 50624 2020-12-31          2953955753 1215668032934117378     1
#> 50625 2020-12-31          2953955753           143689068     1
#> 50628 2020-12-31          2953955753           500964169    12
#> 50629 2020-12-31           297543171            37494271     4
#> 50630 2020-12-31           297543171            77009449     3
#> 50631 2020-12-31           297784596           297784596     3
#> 50634 2020-12-31           299298402           299298402     2
#> 50639 2020-12-31           301464513           301464513     1
#> 50641 2020-12-31           301464513           332993154    17
#> 50670 2020-12-31          3256203501           143852870     4
#> 50700 2020-12-31          3313610230 1132308934948282368     1
#> 50701 2020-12-31          3313610230           134232117     1
#> 50703 2020-12-31          3313610230           148194379     1
#> 50706 2020-12-31          3313610230          1584076484     1
#> 50712 2020-12-31          3313610230          2953955753     1
#> 50718 2020-12-31          3313610230          4449919888     5
#> 50719 2020-12-31          3313610230          4707533116     1
#> 50721 2020-12-31          3313610230           535633686     3
#> 50722 2020-12-31          3313610230  724991849912500224     1
#> 50723 2020-12-31          3313610230             8565932     1
#> 50730 2020-12-31           332993154           301464513     1
#> 50779 2020-12-31            34455479            37494271     2
#> 50780 2020-12-31            34455479            77009449    59
#> 50782 2020-12-31           349968963           349968963     1
#> 50783 2020-12-31           351953986           351953986    67
#> 50784 2020-12-31           351953986          4020276615     1
#> 50791 2020-12-31            35733038           257460186     1
#> 50846 2020-12-31          3763467682 1132308934948282368     1
#> 50847 2020-12-31          3763467682 1208861562410283010     1
#> 50851 2020-12-31          3763467682           185076776     1
#> 50852 2020-12-31          3763467682          2953955753     7
#> 50853 2020-12-31          3763467682          3763467682     2
#> 50855 2020-12-31          3763467682           500964169    17
#> 50856 2020-12-31          3763467682  779340061263425537    55
#> 50857 2020-12-31          3763467682  879024609227665408     1
#> 50880 2020-12-31           398221268            38494089     1
#> 50881 2020-12-31          4020276615 1095773453855137794     1
#> 50882 2020-12-31          4020276615           175273213     3
#> 50883 2020-12-31          4020276615           351953986    23
#> 50884 2020-12-31          4020276615          4020276615    48
#> 50885 2020-12-31          4020276615            56818132     4
#> 50898 2020-12-31            40858089            16492551     1
#> 50900 2020-12-31            40858089            18143118     2
#> 50903 2020-12-31            40858089            48568308     1
#> 50904 2020-12-31            40858089           525264466     1
#> 50905 2020-12-31            40858089            54414081     1
#> 50915 2020-12-31          4198001913           284783298     2
#> 50964 2020-12-31          4449919888           112822128     2
#> 50965 2020-12-31          4449919888 1132308934948282368     2
#> 50967 2020-12-31          4449919888           143689068     1
#> 50970 2020-12-31          4449919888          2841506345     1
#> 50971 2020-12-31          4449919888          2953955753     1
#> 50975 2020-12-31          4449919888  724991849912500224     1
#> 51010 2020-12-31          4707533116           112822128     3
#> 51011 2020-12-31          4707533116 1132308934948282368     1
#> 51013 2020-12-31          4707533116           143689068     2
#> 51016 2020-12-31          4707533116          1579326655     1
#> 51018 2020-12-31          4707533116          2953955753     9
#> 51020 2020-12-31          4707533116           500964169     5
#> 51022 2020-12-31          4707533116  724991849912500224     1
#> 51023 2020-12-31          4707533116  779340061263425537     1
#> 51040 2020-12-31            48568308            24900072     1
#> 51043 2020-12-31           485988630          1578961946     4
#> 51046 2020-12-31           485988630           249976489     1
#> 51071 2020-12-31           500964169 1215668032934117378     1
#> 51072 2020-12-31           500964169           143689068     1
#> 51073 2020-12-31           500964169          1579326655     1
#> 51075 2020-12-31           500964169           185076776     1
#> 51076 2020-12-31           500964169          2953955753     1
#> 51078 2020-12-31           500964169          4707533116     2
#> 51079 2020-12-31           500964169  779340061263425537     2
#> 51111 2020-12-31            54414081            18143118     1
#> 51149 2020-12-31           576528582            24900072     1
#> 51170 2020-12-31           593781032          2953955753     1
#> 51171 2020-12-31           593781032           500964169     4
#> 51201 2020-12-31            63511700 1132308934948282368     6
#> 51206 2020-12-31            63511700           500964169     1
#> 51207 2020-12-31            63511700  724991849912500224     1
#> 51208 2020-12-31            65360466           175422899    15
#> 51209 2020-12-31            65360466           257432060    24
#> 51211 2020-12-31            65360466          3167567897     3
#> 51259 2020-12-31           707334512            48568308     1
#> 51295 2020-12-31            77009449            37494271     5
#> 51297 2020-12-31            79749136            79749136     1
#> 51313 2020-12-31            84713190          4449919888     1
#> 51314 2020-12-31             8565932 1132308934948282368    11
#> 51319 2020-12-31             8565932          2953955753    14
#> 51324 2020-12-31             8565932           500964169     1
#> 51325 2020-12-31             8565932  724991849912500224     6
#> 51458 2020-12-31           950224302          2953955753     2
#> 51459 2020-12-31           950224302           500964169     1
#> 51461 2020-12-31            95970921            18143118     1
#> 51462 2020-12-31            95970921            24900072     4